# [Layer] Improve forwarding logic of ConcatLayer @open sesame 08/08 18:22