## Description

This pull request aims to optimize the performance of `forwarding()` in `ConcatLayer`.
This is accomplished by improving the current concatenation logic by utilizing the tensor operation `concat()` instead of `Map()`.
Note that this optimization is only for `forwarding()` and `calcDeriv()` would be improved in later PR.

## Result

#### Before
|key|avg|min|max|sum|
|------|---|---|---|---|
|calcDeriv(concat)|1.004153s|0.997947s|1.010360s|2.008307s|
|forward(concat)|1.083193s|1.042626s|1.010360s|2.166386s|
|concat|1.129055s|0.997965s|1.252510s|9.032440s|

#### After
|key|avg|min|max|sum|
|------|---|---|---|---|
|calcDeriv(concat)|1.134034s|1.131812s|1.136257s|2.268069s|
|forward(concat)|0.017167s|0.012304s|0.022031s|0.034335s|
|concat|0.627868s|0.012310s|1.346551s|5.022949s|

## Note

input 0: 1:1:196:256 [ FP16 : NCHW ]
input 1: 1:1:196:1024 [ FP16 : NCHW ]
output: 1:1:196:1280 [ FP16 : NCHW ]